moving ErrorResponseDto to error package
